+++ src/Symfony/SymfonyIdentityResolver.php
@@ -10,6 +10,9 @@ use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 use Symfony\Component\Security\Core\User\UserInterface;
 
+/**
+ * @internal
+ */
 final readonly class SymfonyIdentityResolver implements IdentityResolverInterface
 {
     public function __construct(